requirements.txt
asgiref==3.2.8
boto3==1.14.58
botocore==1.17.58
dj-database-url==0.5.0
Django==3.0.7
django-heroku==0.3.1
django-storages==1.10
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.10.0
libmagic==1.0
Pillow==7.1.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
python-dateutil==2.8.1
python-magic
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
XlsxWriter==1.2.9
